pyproject.toml
[tool.poetry]
name = "testopeneats"
version = "0.1.0"
description = ""
authors = ["Philippe Ivan Mbarga <pmbarga@simco.fr>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.2"
djangorestframework = "^3.14.0"
django-filter = "^23.1"
drf-spectacular = "^0.26.2"

[tool.poetry.dev-dependencies]
pre-commit = "^3.2.2"
pytest-django = "^4.5.2"
black = "^23.3.0"
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
